Removed redundant regular expression search in full_gerber_size
The regular expression search for both X and Y at the same time is redundant and can be removed.  The searches for X and Y individually that follow it will catch cases where X and Y are combined.

The regular expression is searching for a substring of the line, not trying to match an entire line, therefore, it does not matter if X is alone or part of X and Y together on a single line.